# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UniAssist API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks/": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create",
        "description": "Professor or TA creates a new task.",
        "operationId": "create_tasks__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "List All",
        "description": "List tasks. Students automatically filter by their cohort + major and if the task is not due.\nStaff can see all tasks or filter manually.",
        "operationId": "list_all_tasks__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cohort_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cohort Year"
            }
          },
          {
            "name": "major",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Major"
            }
          },
          {
            "name": "task_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Type"
            }
          },
          {
            "name": "filter_due_tasks",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter for tasks whose due date has not passed",
              "default": false,
              "title": "Filter Due Tasks"
            },
            "description": "Filter for tasks whose due date has not passed"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskListItemResponse"
                  },
                  "title": "Response List All Tasks  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/submissions": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Task Submissions",
        "operationId": "task_submissions_tasks__task_id__submissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "latest_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Latest Only"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SubmissionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionQueueItem"
                  },
                  "title": "Response Task Submissions Tasks  Task Id  Submissions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get One",
        "description": "Get a single task with its type-specific details.",
        "operationId": "get_one_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete",
        "description": "Professor deletes a task.\nCascades to all related rows: rubrics, submissions, files, chat sessions.\nTAs cannot delete tasks.",
        "operationId": "delete_tasks__task_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovedRubricResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/RubricCriterionInput"
            },
            "type": "array",
            "title": "Criteria"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "total",
          "criteria"
        ],
        "title": "ApprovedRubricResponse",
        "description": "Student-facing grading expectations, never assessment output or staff drafts."
      },
...
      "CreateTaskRequest": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "lab",
              "assignment",
              "project"
            ],
            "title": "Type"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "target_cohort_year": {
            "type": "integer",
            "minimum": 1,
            "title": "Target Cohort Year"
          },
          "target_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Major"
          },
          "reference_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference File Id"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          },
          "allowed_file_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed File Types"
          },
          "allow_late": {
            "type": "boolean",
            "title": "Allow Late",
            "default": false
          },
          "default_repo_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Repo Provider",
            "default": "github"
          },
          "require_team": {
            "type": "boolean",
            "title": "Require Team",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "description",
          "due_date",
          "target_cohort_year"
        ],
        "title": "CreateTaskRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RubricCriterionInput": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "max_points": {
            "type": "number",
            "exclusiveMinimum": 0,
            "title": "Max Points"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "type": "object",
        "required": [
          "name",
          "max_points"
        ],
        "title": "RubricCriterionInput"
      },
...
      "SubmissionEligibility": {
        "properties": {
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "allowed": {
            "type": "boolean",
            "title": "Allowed"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "rubric_ready": {
            "type": "boolean",
            "title": "Rubric Ready"
          },
          "rubric_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Total"
          }
        },
        "type": "object",
        "required": [
          "allowed",
          "rubric_ready"
        ],
        "title": "SubmissionEligibility"
      },
...
      "SubmissionListItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "ai_suggested_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Grade"
          },
          "final_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Grade"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "total_possible_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Possible Grade"
          }
        },
        "type": "object",
        "required": [
          "id",
          "attempt_number",
          "is_latest",
          "status",
          "submitted_at"
        ],
        "title": "SubmissionListItemResponse"
      },
      "SubmissionQueueItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "ai_suggested_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Grade"
          },
          "final_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Grade"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "total_possible_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Possible Grade"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "student_number": {
            "type": "string",
            "title": "Student Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "attempt_number",
          "is_latest",
          "status",
          "submitted_at",
          "student_id",
          "student_name",
          "student_number"
        ],
        "title": "SubmissionQueueItem"
      },
      "SubmissionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "ai_graded",
          "staff_confirmed"
        ],
        "title": "SubmissionStatus"
      },
...
      "TaskAssignmentDetailsResponse": {
        "properties": {
          "allowed_file_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed File Types"
          },
          "allow_late": {
            "type": "boolean",
            "title": "Allow Late",
            "default": false
          }
        },
        "type": "object",
        "title": "TaskAssignmentDetailsResponse"
      },
      "TaskCreatedResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "due_date",
          "created_at"
        ],
        "title": "TaskCreatedResponse"
      },
      "TaskDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "target_cohort_year": {
            "type": "integer",
            "title": "Target Cohort Year"
          },
          "target_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Major"
          },
          "reference_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference File Id"
          },
          "latest_submission": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmissionListItemResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_counts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Counts"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "submission_eligibility": {
            "$ref": "#/components/schemas/SubmissionEligibility"
          },
          "accepted_rubric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovedRubricResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "lab_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskLabDetailsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignment_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskAssignmentDetailsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "project_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskProjectDetailsResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "description",
          "due_date",
          "target_cohort_year",
          "created_by",
          "created_at",
          "submission_eligibility"
        ],
        "title": "TaskDetailResponse"
      },
      "TaskLabDetailsResponse": {
        "properties": {
          "scheduled_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          }
        },
        "type": "object",
        "title": "TaskLabDetailsResponse"
      },
      "TaskListItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "target_cohort_year": {
            "type": "integer",
            "title": "Target Cohort Year"
          },
          "target_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Major"
          },
          "reference_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference File Id"
          },
          "latest_submission": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmissionListItemResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_counts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Counts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "description",
          "due_date",
          "target_cohort_year"
        ],
        "title": "TaskListItemResponse"
      },
      "TaskProjectDetailsResponse": {
        "properties": {
          "default_repo_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Repo Provider"
          },
          "require_team": {
            "type": "boolean",
            "title": "Require Team",
            "default": true
          }
        },
        "type": "object",
        "title": "TaskProjectDetailsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}